% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciFun.R
\name{ciFun}
\alias{ciFun}
\title{Calculates and plots confidence intervals for random effects.}
\usage{
ciFun(model, model.variable, data = NULL, col.variable = NULL, plot = T,
  stat = "median", level = 0.95, oddsRatio = F, nsim = 1000,
  sig.only = FALSE, group.variable, ...)
}
\arguments{
\item{model}{A merMod object from \pkg{lme4}.}

\item{model.variable}{Character. The name of the random effect variable to be computed.}

\item{data}{An optional data frame used in \emph{model}}

\item{col.variable}{An optional variable name used to colour the plotted values}

\item{plot}{Logical. Should a plot be produced?}

\item{stat}{\emph{median} or \emph{mean}}

\item{level}{Numeric. Confidence interval to return}

\item{oddsRatio}{Logical. should paramters be converted to odds ratios?}

\item{nsim}{Numeric.Number of simulations to use}

\item{sig.only}{Logical. Should only significant random effect levels be plotted?}

\item{...}{other arguments passed to REsim.}
}
\value{
A data frame and plot of confidence intervals.
}
\description{
Calculates and plots confidence intervals for random effects of \pkg{lme4} models using \code{\link[lme4]{REsim}}
}
\details{
Use the Gelman sim technique to build empirical Bayes estimates. Uses the sim function in the arm package
}
\examples{
require(lme4)
m1 <- lmer(Reaction ~ 1 + (Days | Subject), sleepstudy)
ci1 <- ciFun(model=m1, model.variable= "Days")

d1<-cbpp
d1$period<-as.numeric(as.character(cbpp$period))
d1$response<-d1$incidence/d1$size
gm1 <- glmer(response ~ 1 + (period | herd), data = d1, weights=d1$size, family = binomial)
ci2 <- ciFun(model=gm1, model.variable= "period")
}

