% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influenceFun.R
\name{influenceFun}
\alias{influenceFun}
\title{Identify highly influential data points.}
\usage{
influenceFun(model, model.variable, data, group.variable,
  inf.type = c("observations", "groups"), plot = T, ...)
}
\arguments{
\item{model}{A merMod object from \pkg{lme4}.}

\item{model.variable}{Character. The name of the random effect variable to be computed.}

\item{data}{An optional data frame used in \emph{model}}

\item{group.variable}{Character. Grouping variable for plots}

\item{inf.type}{Character vector. Specifying which type of influential variables are returned}

\item{plot}{Logical. Should a plot be produced?}

\item{...}{Other arguments passed to \code{\link{ciFun}}}
}
\description{
Use \code{\link[influence.ME]{influence.mer}} from the \pkg{influence.ME} package. Influential observations
and/or groups are identified as being above the critical threshold of \eqn{4/N}, where \eqn{N} is the number of
unique observations or groups.
}
\examples{
require(lme4)
m1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
inf1<-influenceFun(model=m1, model.variable= "Days", data=sleepstudy, group.variable="Subject")
inf1

d1<-cbpp
d1$period<-as.numeric(as.character(cbpp$period))
d1$response<-d1$incidence/d1$size
gm1 <- glmer(cbind(incidence, size - incidence) ~ 1 + (period | herd), data = d1, family = binomial)

inf1<-influenceFun(model=gm1, model.variable= "period", data=d1, group.variable="herd")
inf1

gm2 <- glmer(response ~ 1 + (period | herd), data = d1, weights=d1$size, family = binomial)
inf2<-influenceFun(model=gm2, model.variable= "period", data=d1, group.variable="herd")
inf2
}

