% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelaverageFun.R
\name{modelaverageFun}
\alias{modelaverageFun}
\title{Uses model avaraging to return fixed-effects parameter estimates.}
\usage{
modelaverageFun(model, threshold, fixed = NULL, r2glmm = F)
}
\arguments{
\item{model}{A merMod object from \pkg{lme4}. Most complicated model to evaluate.}

\item{threshold}{Character. \emph{"C95"}, \emph{"D2"}, \emph{"D4"}, \emph{"D6"}, \emph{"D8"}}

\item{fixed}{optional, either a single sided formula or a character vector giving names of terms to be included in all models}

\item{r2glmm}{Logical. Compute R2c and R2m for (g)lmm models using \code{\link[MuMIn]{r.squaredGLMM}}}
}
\value{
\emph{"$all.models"} Object of class \code{\link[MuMIn]{dredge}}

\emph{"$best.models"} Object of class \code{\link[MuMIn]{model.avg}}
}
\description{
Uses model avaraging to return fixed-effects parameter estimates.
}
\details{
Uses the \code{\link[MuMIn]{dredge}} function which takes the full (i.e. most complex, including all terms of interest and interactions) model and automatically
builds all combinations of simpler models from the provided terms. For each model, AIC values are computed. An average model is built \code{\link[MuMIn]{model.avg}}
using the subset of best models determined by the \emph{threshold} argument. \emph{"C95"} will select the models whose cumulative AICw ≤ 0.95. \emph{"D2"}, \emph{"D4"}, \emph{"D6"}, \emph{"D8"}
will select models with ΔAIC ≤ 2, ΔAIC ≤ 4, ΔAIC ≤ 6 and ΔAIC ≤ 8 respectively. If only a single model meets the threshold, it will be returned
instead of the averaged model.
}
\examples{
require(lme4)
d1<-cbpp
d1$period<-as.numeric(as.character(cbpp$period))
d1$response<-d1$incidence/d1$size
gm1 <- glmer(cbind(incidence, size - incidence) ~ period+size + (1 | herd), data = cbpp, family = binomial)
mm1 <- modelaverageFun(model=gm1, threshold="D2")
mm1$all.models
summary(mm1$best.models)
}

