% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplotFun.R
\name{xyplotFun}
\alias{xyplotFun}
\title{A wrapper around \code{\link[lattice]{xyplot}} with showing significance from \code{\link{ciFun}}.}
\usage{
xyplotFun(model, model.variable, data, group.variable, return = "both",
  type = c("p", "r"), ...)
}
\arguments{
\item{model}{A merMod object from \pkg{lme4}.}

\item{model.variable}{Character. The name of the random effect variable to be computed.}

\item{data}{An optional data frame used in \emph{model}}

\item{group.variable}{Character. Grouping variable for plots}

\item{return}{Character. Either \emph{"raw"}, \emph{"fitted"}, or \emph{"both"}}

\item{type}{Character. Passed to \emph{type} in \code{\link[lattice]{xyplot}}}

\item{...}{Other arguments passed to \code{\link{ciFun}}}
}
\description{
Grouping variables that do not overlap 0 are showin with *+* and *-*
}
\examples{
require(lme4)
m1 <- lmer(Reaction ~ 1 + (Days | Subject), sleepstudy)
xyplotFun(model=m1, model.variable= "Days", data=sleepstudy, group.variable="Subject")

d1<-cbpp
d1$period<-as.numeric(as.character(cbpp$period))
d1$response<-d1$incidence/d1$size
gm1 <- glmer(response ~ 1 + (period | herd), data = d1, weights=d1$size, family = binomial)
xyplotFun(model=gm1, model.variable= "period", data=d1, group.variable="herd")
}

